Crunch branch UI
Remove new UI files that breaks build
