+++ selfdrive/ui/SConscript
@@ -22,7 +22,6 @@ if arch == "Darwin":
 widgets_src = ["qt/util.cc",
                "qt/widgets/input.cc", "qt/widgets/drive_stats.cc",
                "qt/widgets/ssh_keys.cc", "qt/widgets/toggle.cc", "qt/widgets/controls.cc",
-               "qt/widgets/branch_control.cc",
                "qt/widgets/offroad_alerts.cc", "qt/widgets/setup.cc", "qt/widgets/keyboard.cc",
                "qt/widgets/scrollview.cc", "qt/widgets/cameraview.cc","qt/widgets/main_sidebar_buttons.cc","qt/widgets/nav_buttons.cc", "#phonelibs/qrcode/QrCode.cc", "qt/api.cc",
                "qt/request_repeater.cc"]
